Gameplay: Fix long note combo count on high density
- Combo count should match arcade now
- Changed particle service to use a circular buffer instead to avoid
spamming too many particles
- Also limtied hold and arc notes to spawn one particle per frame only

+++ Assets/Scripts/Gameplay/Judgement/Input/TouchInputHandler.cs
@@ -215,7 +215,7 @@ namespace ArcCreate.Gameplay.Judgement.Input
 
                     if (input.Phase == UnityEngine.InputSystem.TouchPhase.Ended)
                     {
-                        colorLogic.FingerLifted(input.Id, req.Arc.TimeIncrement);
+                        colorLogic.FingerLifted(input.Id, (float)req.Arc.TimeIncrement);
                         continue;
                     }
