+++ Assets/Scripts/Gameplay/Score/ScoreEvent.cs
@@ -3,6 +3,6 @@ namespace ArcCreate.Gameplay.Score
     public struct ScoreEvent
     {
         public int Timing;
-        public float Score;
+        public double Score;
     }
 }
